requirements.txt
Django==2.2.19
djangorestframework==3.12.4
Pillow==9.4.0
pytz==2022.7.1
sqlparse==0.4.3
